introduce conftest.py
